+++ test/integration/distributed2/serverspec/server_spec.rb
@@ -6,13 +6,13 @@ describe 'gluster::server.rb' do
   end
 
   describe command('gluster volume status') do
-    its(:stdout) { should include('Brick gluster1:/data/sdb1/gv0') }
+    its(:stdout) { should include('Brick gluster1:/data/gv0/brick') }
   end
 
   describe command('gluster volume info') do
     its(:stdout) { should include('Type: Distribute') }
     its(:stdout) { should include('Number of Bricks: 2') }
-    its(:stdout) { should include('gluster2:/data/sdb1/gv0') }
-    its(:stdout) { should include('gluster1:/data/sdb1/gv0') }
+    its(:stdout) { should include('gluster2:/data/gv0/brick') }
+    its(:stdout) { should include('gluster1:/data/gv0/brick') }
   end
 end